Refactor kgof module (#318)
* Created kgof module

* Added tutorial documentation

* Resolved an import statement issue

* first pass of removal of extraneous code

* removed majority of extraneous functions in _utils.py

* Further removal of redundancies + addition of citations

* Removal of redundancies / unnecessary features

* Removed NumpySeedContext to resolve redundancies with numpy.random default_rng

* Removed further dependencies on np.random and replaced with default_rng()

* Removed subsample_ind

* added testing file for kernel.py

* black formatting applied

* added testing file for density.py and replaces all instances of np.random with default_rng() in data.py

* Update index.rst

* Updated circleci so tests run

* Cleaned up docstrings and moved around functions (formatting)

* Lots of cleanup and PR comments addressed

* Rewrote necessary unit tests with pytest

* Black reformatting

* Removed extraneous imports

* Dependencies installation attempt

* Changed dependencies

* removed extraneous imports

* removed dependency on future

* Fixed imports

* Update fssd.py

* fixed circular import issue

* Black reformatting

* Fixed import statements in unit test files

* removed unittest

* Modified import statements

* Made testing edits

* Testing edits

* Black reformatting

* Update test_fssd.py

* Testing edits

* Removed extraneous functions, rewrote tests

* Update test_density.py

* Removed extraneous functions from density.py

* Removed extraneous functions and imports

* Added test for density.Normal

* Added tests for kernel.py

* Fixed import error in test_kernel

* Removed extraneous function from density.py

* Removed extraneous GaussianMixture support

* Added tests for data.py and removed extraneous function from data.py

* Added tests for fssd.py

* Added tests to test_kernel.py

* Made changes to fssd syntax and added more tests for fssd

* Added tests for fssd.py

* black reformatting

* Added more tests for data,py

* density.py and data.py tests added

* Removed extraneous code

* Added more tests for fssd.py

* Formatting fix

* Changed description of base class.

* Added kgof tutorial to index.rst, reformatted docstrings

* Black reformatting in base.py

* Fixed docstrings in fssd.py

* Reformatted kgof tutorial

* Fixed documentation and tutorial rendering issues

* Fixed issues for misc functions not building

* add current module back to misc and base class

* Tutorial formatting changes

* [skip ci] fix nabla math symbol

* Removed helper functions from fssd class file

* Made changes to test new modified fssd.py

* Removed Data object inputs and replaced them with numpy array inputs

* Removed Data dependent tests

* Continued edits to testing files following Data removal

* Removed redundant code, used compute_kern from tools

* Updated kernel testing file to evaluate usage of tools functions

* Added scipy.stats functionality to density sims

* Update kernel.py

* Removed extraneous references to Data

* Black reformatting

* Black reformatting

* Made edits to the testing files to run more robust tests

* black reformatting

* Installed newest black build and reformatted files

* Update config.yml

* Resolved circular import issue

* fixed import statement

* Delete data.py

* Fixed unexpected argument issue

* Removed extraneous testing code

* fixed autograd ArrayBox TypeError

* Testing ArrayBox issue

* Reformatting

* Update test_density.py

* Update test_density.py

* Edits based on comments

* Addressed comments and made testing changes

* Raised ValueErrors, testing changes

* Update test_density.py

* Made small testing edits

* Updates to datasources

* Small testing edits

* Update test_density.py

* Update test_density.py

* testing edits

* Update test_density.py

* Density + datasource testing edits

* Update test_density.py

* Update test_density.py

* Update test_density.py

* Update test_density.py

* Update test_density.py

* Removed extraneous function

* Testing edit

* Updated index.rst based on instruction

* Minor changes to index.rst and kernel testing addressed

* Added T0D0 for removal of kernel.py

* fix spaces in index.rst
