+++ docs/index.rst
@@ -2,7 +2,7 @@
 
 .. include:: ../README.md
     :parser: myst_parser.sphinx_
-
+    
 .. toctree::
     :hidden:
     :caption: Getting Started
