+++ hyppo/kgof/__init__.py
@@ -1,6 +1,6 @@
 from ._utils import meddistance, fit_gaussian_draw
 from .fssd import FSSD, FSSDH0SimCovObs
-from .data import Data
+from .datasource import DataSource
 from .density import Normal, IsotropicNormal
 from .kernel import KGauss
