+++ hyppo/kgof/fssd.py
@@ -33,7 +33,7 @@ class H0Simulator(ABC):
         self.seed = seed
 
     @abstractmethod
-    def simulate(self, gof, dat):
+    def simulate(self, gof, X):
         """
         gof: a GofTest
         dat: a Data (observed data)
@@ -59,7 +59,7 @@ class FSSDH0SimCovObs(H0Simulator):
     def __init__(self, n_simulate=3000, seed=10):
         super(FSSDH0SimCovObs, self).__init__(n_simulate, seed)
 
-    def simulate(self, gof, dat, fea_tensor=None):
+    def simulate(self, gof, X, fea_tensor=None):
         """
         fea_tensor: n x d x J feature matrix
         """
@@ -67,17 +67,15 @@ class FSSDH0SimCovObs(H0Simulator):
         n_simulate = self.n_simulate
         seed = self.seed
         if fea_tensor is None:
-            _, fea_tensor = gof.statistic(dat, return_feature_tensor=True)
+            _, fea_tensor = gof.statistic(X, return_feature_tensor=True)
 
         J = fea_tensor.shape[2]
-        X = dat.data()
         n = X.shape[0]
         # n x d*J
         Tau = fea_tensor.reshape(n, -1)
         # Make sure it is a matrix i.e, np.cov returns a scalar when Tau is
         # 1d.
         cov = np.cov(Tau.T) + np.zeros((1, 1))
-        # cov = Tau.T.dot(Tau/n)
 
         arr_nfssd, eigs = list_simulate_spectral(cov, J, n_simulate, seed=self.seed)
         return {"sim_stats": arr_nfssd}
@@ -102,14 +100,13 @@ class FSSDH0SimCovDraw(H0Simulator):
         super(FSSDH0SimCovDraw, self).__init__(n_simulate, seed)
         self.n_draw = n_draw
 
-    def simulate(self, gof, dat, fea_tensor=None):
+    def simulate(self, gof, fea_tensor=None):
         """
         fea_tensor: n x d x J feature matrix
         This method does not use dat.
 
         From: https://github.com/wittawatj/fsic-test
         """
-        dat = None
         # p = an UnnormalizedDensity
         p = gof.p
         ds = p.get_datasource()
@@ -118,7 +115,7 @@ class FSSDH0SimCovDraw(H0Simulator):
         Xdraw = ds.sample(n=self.n_draw, seed=self.seed)
         _, fea_tensor = gof.statistic(Xdraw, return_feature_tensor=True)
 
-        X = Xdraw.data()
+        X = Xdraw
         J = fea_tensor.shape[2]
         n = self.n_draw
         # n x d*J
@@ -201,7 +198,7 @@ class FSSD(GofTest):
         self.V = V
         self.null_sim = null_sim
 
-    def test(self, dat, return_simulated_stats=False):
+    def test(self, X, return_simulated_stats=False):
         r"""
         Perform the goodness-of-fit test using an FSSD test statistic
         and return values computed in a dictionary.
@@ -218,12 +215,11 @@ class FSSD(GofTest):
         alpha = self.alpha
         null_sim = self.null_sim
         n_simulate = null_sim.n_simulate
-        X = dat.data()
         n = X.shape[0]
         J = self.V.shape[0]
 
-        nfssd, fea_tensor = self.statistic(dat, return_feature_tensor=True)
-        sim_results = null_sim.simulate(self, dat, fea_tensor)
+        nfssd, fea_tensor = self.statistic(X, return_feature_tensor=True)
+        sim_results = null_sim.simulate(self, X, fea_tensor)
         arr_nfssd = sim_results["sim_stats"]
 
         # approximate p-value with the permutations
@@ -240,7 +236,7 @@ class FSSD(GofTest):
             results["sim_stats"] = arr_nfssd
         return results
 
-    def statistic(self, dat, return_feature_tensor=False):
+    def statistic(self, X, return_feature_tensor=False):
         r"""
         Compute the test statistic. The statistic is n*FSSD^2.
 
@@ -252,7 +248,6 @@ class FSSD(GofTest):
         -------
         stat : the test statistic, n*FSSD^2
         """
-        X = dat.data()
         n = X.shape[0]
 
         # n x d x J
@@ -265,7 +260,7 @@ class FSSD(GofTest):
         else:
             return stat
 
-    def get_H1_mean_variance(self, dat):
+    def get_H1_mean_variance(self, X):
         r"""
         Calculate the mean and variance under H1 of the test statistic (divided by
         n).
@@ -279,7 +274,6 @@ class FSSD(GofTest):
         mean : the mean under the alternative hypothesis of the test statistic
         variance : the variance under the alternative hypothesis of the test statistic
         """
-        X = dat.data()
         Xi = self.feature_tensor(X)
         mean, variance = ustat_h1_mean_variance(Xi, return_variance=True)
         return mean, variance
@@ -319,9 +313,7 @@ class FSSD(GofTest):
         return Xi
 
 
-def power_criterion(
-    p, dat, k, test_locs, reg=1e-2, use_unbiased=True, use_2terms=False
-):
+def power_criterion(p, X, k, test_locs, reg=1e-2, use_unbiased=True, use_2terms=False):
     r"""
     Compute the mean and standard deviation of the statistic under H1.
 
@@ -337,7 +329,6 @@ def power_criterion(
     -------
     obj : mean/sd
     """
-    X = dat.data()
     n = X.shape[0]
     V = test_locs
     fssd = FSSD(p, k, V, null_sim=None)
@@ -453,7 +444,7 @@ def simulate_null_dist(eigs, J, n_simulate=2000, seed=7):
     return fssds
 
 
-def fssd_grid_search_kernel(p, dat, test_locs, list_kernel):
+def fssd_grid_search_kernel(p, X, test_locs, list_kernel):
     r"""
     Linear search for the best kernel in the list that maximizes
     the test power criterion, fixing the test locations to V.
@@ -470,12 +461,11 @@ def fssd_grid_search_kernel(p, dat, test_locs, list_kernel):
     objs : array of test power criteria
     """
     V = test_locs
-    X = dat.data()
     n_cand = len(list_kernel)
     objs = np.zeros(n_cand)
     for i in range(n_cand):
         ki = list_kernel[i]
-        objs[i] = power_criterion(p, dat, ki, test_locs)
+        objs[i] = power_criterion(p, X, ki, test_locs)
         logging.info("(%d), obj: %5.4g, k: %s" % (i, objs[i], str(ki)))
 
     # Widths that come early in the list
