+++ hyppo/kgof/tests/test_density.py
@@ -3,7 +3,7 @@ from numpy import testing
 from past.utils import old_div
 from scipy.linalg.misc import norm
 
-from ..data import DSNormal
+from ..datasource import DSNormal, DSIsotropicNormal
 from ..density import IsotropicNormal, Normal
 import scipy.stats as stats
 from numpy.random import default_rng
@@ -21,6 +21,7 @@ class TestIsotropicNormal:
         X = rng.random(size=(n, d)) + 1
 
         isonorm = IsotropicNormal(mean, variance)
+        ds_isonorm = DSIsotropicNormal(mean, variance)
         isonorm.log_normalized_den(X)
         isonorm.dim()
         log_dens = isonorm.log_den(X)
@@ -67,7 +68,6 @@ class TestNormal:
 
         ds_norm = DSNormal(test_mean, cov)
         ds_norm.sample(n=10)
-        ds_norm.dim()
         norm.get_datasource()
         norm.dim()
