+++ hyppo/kgof/tests/test_fssd.py
@@ -3,15 +3,14 @@ import numpy.testing as testing
 
 from ..fssd import (
     FSSD,
-    FSSDH0SimCovObs,
-    FSSDH0SimCovDraw,
     ustat_h1_mean_variance,
     power_criterion,
     fssd_grid_search_kernel,
+    FSSDH0SimCovObs,
+    FSSDH0SimCovDraw,
 )
 from .._utils import meddistance, fit_gaussian_draw, constrain
 from ..kernel import KGauss
-from ..data import Data
 from ..density import IsotropicNormal
 
 from numpy.random import default_rng
@@ -38,13 +37,6 @@ class TestFSSD:
 
         rng = default_rng(seed)
         X = rng.standard_normal(size=(n, d)) * np.sqrt(draw_variance) + draw_mean
-        dat = Data(X)
-        dat.dim()
-        dat.sample_size()
-        dat.n()
-        dat.split_tr_te()
-        dat.subsample(n=4)
-        dat.clone()
 
         # Test
         sig2 = meddistance(X, subsample=1000) ** 2
@@ -56,13 +48,12 @@ class TestFSSD:
         null_sim = FSSDH0SimCovObs(n_simulate=200, seed=3)
         extra_sim = FSSDH0SimCovDraw()
         fssd = FSSD(isonorm, k, V, null_sim=null_sim, alpha=alpha)
-        check_sim = extra_sim.simulate(dat=dat, gof=fssd)
-        power_criterion(p=isonorm, dat=dat, k=k, test_locs=V)
-        fssd_grid_search_kernel(p=isonorm, dat=dat, test_locs=V, list_kernel=list_k)
-        fssd.get_H1_mean_variance(dat=dat)
+        check_sim = extra_sim.simulate(gof=fssd)
+        power_criterion(p=isonorm, X=X, k=k, test_locs=V)
+        fssd_grid_search_kernel(p=isonorm, X=X, test_locs=V, list_kernel=list_k)
+        fssd.get_H1_mean_variance(X=X)
 
-        tresult = fssd.test(dat, return_simulated_stats=True)
-        dat.__add__(dat)
+        tresult = fssd.test(X, return_simulated_stats=True)
 
         # assertions
         testing.assert_almost_equal(tresult["pvalue"], 0, decimal=1)
@@ -84,11 +75,10 @@ class TestFSSD:
         draw_variance = variance + 1
         rng = default_rng(seed)
         X = rng.standard_normal(size=(n, d)) * np.sqrt(draw_variance) + draw_mean
-        dat = Data(X)
 
         # Test
-        sig2 = meddistance(X, subsample=1000) ** 2
-        k = KGauss(sig2)
+        sig_square = meddistance(X, subsample=1000) ** 2
+        k = KGauss(sig_square)
 
         # random test locations
         V = fit_gaussian_draw(X, J, seed=seed + 1)
@@ -98,6 +88,9 @@ class TestFSSD:
         fea_tensor = fssd.feature_tensor(X)
 
         u_mean, u_variance = ustat_h1_mean_variance(fea_tensor)
+
+        random_sim = FSSDH0SimCovObs()
+        random_sim.simulate(gof=fssd, X=X)
 
         # assertions
         testing.assert_almost_equal(u_variance, 1, decimal=1)
