+++ hyppo/kgof/tests/test_kernel.py
@@ -1,6 +1,7 @@
 import autograd.numpy as np
 from numpy import testing
 from ..kernel import KGauss
+from ...tools import compute_kern, multi_compute_kern
 
 import pytest
 from numpy.random import default_rng
@@ -14,7 +15,7 @@ class TestKGauss:
         rng = default_rng(29)
         X = rng.standard_normal(size=(n, d)) * 3
         k = KGauss(sigma2=1)
-        K = k.eval(X, X)
+        K = k.eval(X, X)  # tests usage of compute_kern
 
         testing.assert_almost_equal(K.shape, (n, n))
 
@@ -37,10 +38,10 @@ class TestKGauss:
         k.pair_eval(X, Y)
         # check correctness
         K = k.eval(X, y[np.newaxis, :])
-        myG = -K / sigma2 * (X - y)
+        modG = -K / sigma2 * (X - y)
 
-        testing.assert_equal(G.shape, myG.shape)
-        testing.assert_almost_equal(G, myG)
+        testing.assert_equal(G.shape, modG.shape)
+        testing.assert_almost_equal(G, modG)
 
     @pytest.mark.parametrize("n", [11])
     @pytest.mark.parametrize("d", [3, 1])
@@ -51,16 +52,15 @@ class TestKGauss:
         k = KGauss(sigma2=sigma2)
 
         # n x n
-        myG = np.zeros((n, n))
+        modG = np.zeros((n, n))
         K = k.eval(X, X)
         for i in range(n):
             for j in range(n):
                 diffi2 = np.sum((X[i, :] - X[j, :]) ** 2)
-                # myG[i, j] = -diffi2*K[i, j]/(sigma2**2)+ d*K[i, j]/sigma2
-                myG[i, j] = K[i, j] / sigma2 * (d - diffi2 / sigma2)
+                modG[i, j] = K[i, j] / sigma2 * (d - diffi2 / sigma2)
 
-        # check correctness
+        # check correctness of function that calculates sum of the gradient across x_i and y_i in Y
         G = k.gradXY_sum(X, X)
 
-        testing.assert_equal(G.shape, myG.shape)
-        testing.assert_almost_equal(G, myG)
+        testing.assert_equal(G.shape, modG.shape)
+        testing.assert_almost_equal(G, modG)
